# Implement ShortestPathManyToMany