--- a/dijkstra_bidirectional.go
+++ b/dijkstra_bidirectional.go
@@ -5,6 +5,14 @@ import (
 	"container/list"
 )
 
+type direction int
+
+const (
+	forward direction = iota
+	backward
+	directionsCount
+)
+
 // ShortestPath Computes and returns shortest path and it's cost (extended Dijkstra's algorithm)
 //
 // If there are some errors then function returns '-1.0' as cost and nil as shortest path