This implements many-to-many simultaneous shortest path algorithm. Note: this allocates data structures `O((sources + targets)*num_vertices)` so the memory usage could be significant. Experimental results show 2-3x performance improvement over iterating `ShortestPathOneToMany` over all sources.

This also changes unifies the forward and backwards searches to reduce code duplication, allowing the `ShortestPathManyToMany` implementation to be simpler.

Also, this PR is currently based on #26 (commit [9e9cd42](https://github.com/LdDl/ch/pull/26/commits/9e9cd42cfbe0bae8c2fa2b756c26a741f8ef0826)) - I could rebase this to current state of master if #26 takes time to merge.