@LdDl 2022-09-15
We'll take a look on this soon (we can't dig into instantly sadly). But we see `go.mod` changes which seems strange.

P.S. BTW #26 has been merged