@magnushiie 2022-09-15
Sorry about the `go.mod` changes, I actually meant to create a PR in our fork first, but accidentally submitted upstream. But maybe you can start evaluating, I will submit an updated patch (rebased on master and without the unnecessary changes) tomorrow or over the weekend.