@magnushiie 2022-09-16
I removed the `go.mod` change but for the rest, let's pause this PR for a while, I'll create separate PRs (e.g. #31, #32) for the smaller changes.