@magnushiie 2022-09-19
I will create one more separate PR tomorrow on top of master branch, to unify the directions logic in OneToOne, and then rebase this one on top of that (it's basically a copy of OneToOne in that state with additional arrays for sources and targets).