@magnushiie 2022-09-20
This is now rebased on top of master, it's based on the restructured OneToOne algorithm from #33. #33 is not necessary for the ManyToMany implementation, this PR can be merged independently. @LdDl if you could please review now, I'd be grateful.