@LdDl 2022-09-21
Did merge.
I've added a couple of benchs
```bash
go test -benchmem -run=^$ github.com/LdDl/ch -bench BenchmarkOldWayShortestPathManyToMany
go test -benchmem -run=^$ github.com/LdDl/ch -bench BenchmarkShortestPathManyToMany
```

`benchcmp`:
```
benchmark                                                                                           old ns/op     new ns/op     delta
BenchmarkShortestPathManyToMany/CH_shortest_path/256/vertices-256-edges-97227-shortcuts-5276-12     1268879       5410451       +326.40%
```

`benchstat`
```
name                                                                                    old time/op    new time/op    delta
ShortestPathManyToMany/CH_shortest_path/256/vertices-256-edges-97227-shortcuts-5276-12    1.27ms ± 0%    5.41ms ± 0%   ~     (p=1.000 n=1+1)
```