@magnushiie 2022-09-21
Thank you!

The current benchmarks don't highlight the benefits of ManyToMany, as there are few sources and few targets. In our real cases (100x100 matrix) this has 2-3x improvement in latency (albeit with increased memory usage).

I added the current test only for code coverage, I will try to add a benchmark with a bigger matrix in a subsequent PR.

Also I forgot about the TODO item https://github.com/LdDl/ch/pull/30/files#diff-be0728d7a21b2446cc27d05a24101ab1250d000630961e717abd16f00002ae29R137 - fixing this could bring the performance to more on par with the other implementations with small cases, and also improve the big matrices. I will submit a PR next week.